+++ esbuild.js
@@ -1,0 +1,15 @@
+const esbuild = require("esbuild");
+
+esbuild
+  .build({
+    entryPoints: ["src/extension.ts"],
+    bundle: true,
+    outfile: "dist/extension.js",
+    external: ["vscode"],
+    format: "cjs",
+    platform: "node",
+    target: "node20",
+    sourcemap: true,
+    minify: false,
+  })
+  .catch(() => process.exit(1));